WIP: added requested changes and small API changes for immutability
